+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Enhancements:**
 
+* Added Steam as authorization option, #311
 * Added likes/dislikes, #322
 * Added new option to control type of likes to display: just likes or likes/dislikes, #322
 * JS main dependency is not much lighter ~600kb